Separate ABI fetching to its own hook, reuse types
