+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+  "semi": true,
+  "tabWidth": 2,
+  "printWidth": 80,
+  "useTabs": false,
+  "singleQuote": false,
+  "trailingComma": "es5"
+}
