+++ src/lib/wagmi.ts
@@ -3,6 +3,7 @@ import { cookieStorage, createConfig, createStorage } from "wagmi";
 import { arbitrum, base, mainnet, optimism, polygon } from "wagmi/chains";
 import { coinbaseWallet, injected, walletConnect } from "wagmi/connectors";
 import { getTransportByChainId } from "./utils";
+import { SupportedChainId } from "@/types";
 
 export const chains = [base, optimism, arbitrum, polygon, mainnet] as const;
 
@@ -16,7 +17,7 @@ export const blockscoutUrlByChainId = {
 
 const transports = Object.fromEntries(
   chains.map((chain) => [chain.id, getTransportByChainId(chain.id)])
-) as { [K in (typeof chains)[number]["id"]]: HttpTransport };
+) as { [K in SupportedChainId]: HttpTransport };
 
 export function getConfig() {
   return createConfig({
