+++ src/types.ts
@@ -1,0 +1,6 @@
+// Only for types that are re-used in multiple files
+
+import { chains } from "./lib/wagmi";
+
+export type SupportedChain = (typeof chains)[number];
+export type SupportedChainId = SupportedChain["id"];
